+++ src/main/java/org/gudy/azureus2/core3/xml/util/XUXmlWriter.java
@@ -185,7 +185,7 @@ XUXmlWriter
 		
 		for ( int i=0;i<chars.length;i++){
 			
-			int	c = (int)chars[i];
+			int	c = chars[i];
 			
 			if (	( c <= 31 ) ||
 					( c >= 127 && c <= 159 ) ||
